take the duration in configuration into account

+++ src/main/java/com/airgap/airgapagent/utils/visitor/PersistentStateListener.java
@@ -23,7 +23,7 @@ public class PersistentStateListener<T> extends SearchEventAdapter<T> {
 
     private StateStore<T> stateStore;
 
-    public PersistentStateListener(int interval, File stateLocation, Serializer<T> converter) {
+    public PersistentStateListener(long interval, File stateLocation, Serializer<T> converter) {
         runner = IntervalRunner.of(Duration.ofSeconds(interval), true);
         this.stateLocation = stateLocation;
         this.converter = converter;
